\name{withoutExclusions-package}
\alias{withoutExclusions-package}
\alias{withoutExclusions}
\docType{package}
\title{
\packageTitle{withoutExclusions}
}
\description{
\packageDescription{withoutExclusions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{withoutExclusions}
\packageIndices{withoutExclusions}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{withoutExclusions}

Maintainer: \packageMaintainer{withoutExclusions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
